+++ components/Layout/Layout.tsx
@@ -1,14 +1,15 @@
 import { ReactNode } from 'react'
 import Head from 'next/head'
 
-import Nav from '../Navigation'
+import { Sidebar, Topbar } from '../Navigation'
 
 interface Props {
     children?: ReactNode
-    title?: string
+    title?: string,
+    darkMode: boolean
 }
 
-const Layout = ({ children, title = 'Kate&John' }: Props) => (
+const Layout = ({ children, title = 'Kate&John', darkMode }: Props) => (
     <>
         <Head>
             <title>{title}</title>
@@ -16,12 +17,17 @@ const Layout = ({ children, title = 'Kate&John' }: Props) => (
             <meta name='viewport' content='initial-scale=1.0, width=device-width' />
             <link rel="icon" href="/favicon.ico" />
         </Head>
-        <header>
-            <Nav />
-        </header>
-        <main className='p-3'>
-            {children}
-        </main>
+        <div className='w-full h-screen flex'>
+            <section className='text-white bg-teal-900 shadow-xl'>
+                <Sidebar />
+            </section>
+            <main className='w-full overflow-x-auto bg-coolGray-100'>
+                <header className='sticky top-0 bg-white shadow-lg'>
+                    <Topbar darkMode={darkMode}/>
+                </header>
+                {children}
+            </main>
+        </div>
     </>
 )