+++ components/Navigation/index.ts
@@ -1,3 +1,2 @@
-import Nav from './Navigation'
-
-export default Nav
+export {default as Sidebar} from './Sidebar'
+export {default as Topbar} from './Topbar'