+++ pages/concerts.tsx
@@ -1,14 +1,11 @@
 import useUser from '../utils/useUser'
-import Layout from '../components/Layout'
 
 const Concerts = () => {
     const { user } = useUser({ redirectTo: '/' })
-    if(!user || !user.isLoggedIn) return <Layout>Loading...</Layout>
+    if(!user || !user.isLoggedIn) return <div>Loading...</div>
 
     return (
-        <Layout>
-            <h1>Concerts</h1>
-        </Layout>
+        <h1>Concerts</h1>
     )
 }